Update frame to use new quantity tag wherever appropriate ...
This is a natural thing to do since it simplifies both the schemas and
the library implementation. The one downside is that default units are
no longer possible for fields that are quantities, but unless there is a
compelling reason to retain that feature, this seems like a good change
to make.