Update wcs schemas to reflect astropy changes (#142)
* Update wcs schemas to reflect astropy changes ...

Astropy changed its representation of the obsgeoloc and obsgeovel
properties from Quantity objects to CartesianRepresentation objects.
This commit adds a new schema and updates the wcs frame schema to
reflect these changes.

* Update frame to use new quantity tag wherever appropriate:

This is a natural thing to do since it simplifies both the schemas and
the library implementation. The one downside is that default units are
no longer possible for fields that are quantities, but unless there is a
compelling reason to retain that feature, this seems like a good change
to make.

* Update frame to use galcen_coord, add ICRS coord schema:

Astropy elimated the fields galcen_ra and galcen_dec for galactocentric
frames and replaced them with galcen_coord, which is an ICRS coordinate
object. This commit brings the ASDF schema in line with these changes

* Add galcen_v_sun property to frame schema:

Astropy added the galcen_v_sun field to Galactocentric frame objects.
This commit updates the ASDF schema to reflect this change.