Fix examples in label_mapper schemas...
These seem to have been malformed for quite some time, but it's not
clear exactly why the tests were passing anyway. This issue was
discovered in the course of updating the asdf package to use ruamel.yaml
instead of pyyaml. It appears that ruamel.yaml provides proper support
for omap whereas pyyaml did not.